+++ src/commands/repository.rs
@@ -1,7 +1,9 @@
 // src/commands/repository.rs
 use crate::cli::RepositoryAction;
 use crate::config::{create_agent, get_config_file, get_sync_dir, get_toml_config};
-use std::fs::{remove_file, write};
+use crate::security::get_file_hash;
+use std::fs::{read_to_string, remove_file, write};
+use std::time::SystemTime;
 use ureq::Agent;
 
 pub fn run(action: RepositoryAction) -> Result<(), Box<dyn std::error::Error>> {
@@ -11,6 +13,36 @@ pub fn run(action: RepositoryAction) -> Result<(), Box<dyn std::error::Error>> {
         RepositoryAction::List => list(),
         RepositoryAction::Sync { name, timeout } => sync(name, timeout),
         RepositoryAction::Check { timeout } => check(timeout),
+        RepositoryAction::Enable { name } => enable_repository(name),
+        RepositoryAction::Disable { name } => disable_repository(name),
+    }
+}
+
+pub enum RepositoryType {
+    Reward,
+    Nested,
+    Unknown,
+}
+
+pub struct Repository {
+    pub repo_type: RepositoryType,
+    pub url: Option<String>,
+}
+
+pub fn parse_repository(entry: String) -> Repository {
+    match entry.splitn(2, ' ').collect::<Vec<_>>().as_slice() {
+        [url] => Repository {
+            repo_type: RepositoryType::Reward,
+            url: Some(url.to_string()),
+        },
+        ["Nested", url] => Repository {
+            repo_type: RepositoryType::Nested,
+            url: Some(url.to_string()),
+        },
+        _ => Repository {
+            repo_type: RepositoryType::Unknown,
+            url: None,
+        },
     }
 }
 
@@ -167,20 +199,33 @@ fn sync(names: Vec<String>, timeout: Option<u64>) -> Result<(), Box<dyn std::err
     parameters:
         - timeout: timeout to check in seconds
     it will iterate all repositories and then check
-    if it is alive or dead
+    if it is alive or dead.
+    It also will show status, last sync time, entry data,
+    hash (sha-256) and errors
 */
 fn check(timeout: Option<u64>) -> Result<(), Box<dyn std::error::Error>> {
     let doc: toml::Value = get_toml_config();
     let mut alive = 0;
     let mut dead = 0;
+    let mut greward = 0;
+    let mut gnested = 0;
+    let mut gunknown = 0;
+    let mut gentry = 0;
+    let mut genabled = 0;
+    let mut gdisabled = 0;
     let agent = create_agent(timeout);
 
     for (name, data) in doc["repositories"].as_table().unwrap() {
-        if !data["enabled"].as_bool().unwrap() {
-            continue;
-        }
-        let url = data["url"].as_str().unwrap().to_string();
         println!("Checking: {}", name);
+        if !data["enabled"].as_bool().unwrap() {
+            println!("  Status: disabled");
+            gdisabled += 1;
+        } else {
+            println!("  Status: enabled");
+            genabled += 1;
+        }
+
+        let url = data["url"].as_str().unwrap().to_string();
         match agent.get(&url).call() {
             Ok(_) => {
                 println!("  Alive");
@@ -191,8 +236,96 @@ fn check(timeout: Option<u64>) -> Result<(), Box<dyn std::error::Error>> {
                 dead += 1;
             }
         }
+
+        match std::fs::metadata(get_sync_dir().join(name)) {
+            Ok(metadata) => {
+                let modified = metadata.modified()?;
+                let elapsed = SystemTime::now().duration_since(modified)?.as_secs();
+
+                println!(
+                    "  Last synced: {} ago",
+                    humantime::format_duration(std::time::Duration::from_secs(elapsed))
+                );
+            }
+            Err(_) => {
+                println!("  Last synced: Never synced");
+            }
+        }
+        let content = read_to_string(get_sync_dir().join(name))?;
+        let entries: Vec<&str> = content
+            .lines()
+            .filter(|l| !l.starts_with('#') && !l.trim().is_empty())
+            .collect();
+
+        let mut reward = 0;
+        let mut nested = 0;
+        let mut unknown = 0;
+        for entry in &entries {
+            let repo_entry = parse_repository(entry.to_string());
+            match repo_entry.repo_type {
+                RepositoryType::Reward => {
+                    reward += 1;
+                }
+                RepositoryType::Nested => {
+                    nested += 1;
+                }
+                RepositoryType::Unknown => {
+                    unknown += 1;
+                }
+            }
+        }
+        println!("  Entry: {}", entries.len());
+        println!("      Reward: {}", reward);
+        println!("      Nested: {}", nested);
+        println!("      Unknown: {}", unknown);
+        println!(
+            "  Hash: {}",
+            get_file_hash(get_sync_dir().join(name).to_str().unwrap()).unwrap()
+        );
+        greward += reward;
+        gnested += nested;
+        gunknown += unknown;
+        gentry += entries.len();
     }
 
-    println!("Checking done:\n Alive: {}\n Dead: {}", alive, dead);
+    println!("Check report:");
+    println!("  Alive: {}", alive);
+    println!("  Dead: {}", dead);
+    println!("  Entry: {}", gentry);
+    println!("    Reward: {}", greward);
+    println!("    Nested: {}", gnested);
+    println!("    Unknown: {}", gunknown);
+    println!("  Enabled: {}", genabled);
+    println!("  Disabled: {}", gdisabled);
+    Ok(())
+}
+
+/*
+    function handler for subcommand enable
+    parameter:
+        - name: name of the repository
+    it will toggle enabled to the repository selected
+*/
+fn enable_repository(name: String) -> Result<(), Box<dyn std::error::Error>> {
+    let mut doc: toml::Value = get_toml_config();
+
+    doc["repositories"][name]["enabled"] = toml::Value::Boolean(true);
+
+    write(get_config_file(), toml::to_string(&doc)?)?;
+    Ok(())
+}
+
+/*
+    function handler for subcommand disable
+    parameter:
+        - name: name of the repository
+    it will toggle disabled to the repository selected
+*/
+fn disable_repository(name: String) -> Result<(), Box<dyn std::error::Error>> {
+    let mut doc: toml::Value = get_toml_config();
+
+    doc["repositories"][name]["enabled"] = toml::Value::Boolean(false);
+
+    write(get_config_file(), toml::to_string(&doc)?)?;
     Ok(())
 }